requirements.txt
# Core
streamlit>=1.30.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Auth
firebase-admin>=6.4.0

# Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# UI enhancements
streamlit-calendar>=0.5.0
plotly>=5.18.0
